+++ source/extensions/isaacsim.asset.importer.mjcf/plugins/isaacsim.asset.importer.mjcf/Platform.cpp
@@ -13,9 +13,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <isaacsim/asset/importer/mjcf/core/Platform.h>
-#include <isaacsim/asset/importer/mjcf/math/core/Core.h>
-#include <isaacsim/asset/importer/mjcf/math/core/Types.h>
+#include <isaacsim/core/includes/core/Platform.h>
+#include <isaacsim/core/includes/math/core/Core.h>
+#include <isaacsim/core/includes/math/core/Types.h>
 
 #include <algorithm>
 #include <ctype.h>