+++ source/extensions/isaacsim.core.experimental.objects/python/tests/test_mesh.py
@@ -21,7 +21,7 @@ import omni.kit.test
 import omni.usd
 import warp as wp
 from isaacsim.core.experimental.objects import Mesh
-from isaacsim.core.experimental.prims.tests.utils import (
+from isaacsim.core.experimental.prims.tests.common import (
     check_allclose,
     check_array,
     check_lists,
@@ -31,7 +31,7 @@ from isaacsim.core.experimental.prims.tests.utils import (
 )
 from pxr import UsdGeom
 
-from .utils import parametrize
+from .common import parametrize
 
 
 async def populate_stage(max_num_prims: int, operation: Literal["wrap", "create"]) -> None: