+++ source/extensions/isaacsim.replicator.examples/python/tests/test_sdg_timeline_running.py
@@ -20,7 +20,7 @@ import omni.replicator.core as rep
 import omni.timeline
 import omni.usd
 
-from .utils import validate_folder_contents
+from .common import validate_folder_contents
 
 
 class TestSDGUsefulSnippets(omni.kit.test.AsyncTestCase):