+++ source/extensions/isaacsim.replicator.examples/python/tests/test_sdg_ur10_palletizing.py
@@ -16,7 +16,7 @@
 import omni.kit
 import omni.usd
 
-from .utils import validate_folder_contents
+from .common import validate_folder_contents
 
 
 class TestSDGUR10Palletizing(omni.kit.test.AsyncTestCase):