+++ source/extensions/isaacsim.asset.importer.mjcf/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [2.4.7] - 2025-05-23
+### Changed
+- Refactor headers into isaacsim.core.includes
+- Add docstrings
+
 ## [2.4.6] - 2025-05-19
 ### Changed
 - Update copyright and license to apache v2.0